// src/api-client/models/color-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ColorRequest
 */
export interface ColorRequest {
    /**
     * 
     * @type {string}
     * @memberof ColorRequest
     */
    'name': string;
}